requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Selenium and browser automation
selenium==4.15.2
undetected-chromedriver==3.5.4
selenium-stealth==1.0.6
webdriver-manager==4.0.1

# Additional utilities
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# Logging and monitoring
loguru==0.7.2

# Optional: For async operations
aiofiles==23.2.0
asyncio-mqtt==0.13.0
